🚧 (wip) extract state from map charts
